+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - Buttons: minor playback controls preset changes.
 - Helpers: support for long paths (>260 chars) in multiple internal file handling functions.
 - Readmes: general cleanup.
+- Readmes: readmes submenus are now renamed to 'Help' in all instances.
 ### Removed
 ### Fixed
 - Playlist Tools\Pools: fix to one of the default presets.